+++ leap_net/keras/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, RTE (https://www.rte-france.com)
+# Copyright (c) 2024, RTE (https://www.rte-france.com)
 # See AUTHORS.txt
 # This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
 # If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
@@ -6,8 +6,16 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of leap_net, leap_net a keras implementation of the LEAP Net model.
 
-# from leap_net.keras_core.ltau import Ltau
-from leap_net.keras_core.ltauNoAdd import LtauNoAdd
+from packaging import version
+
+import keras
+if version.parse(keras.__version__) < version.parse("3"):
+    raise ImportError(f"Impossible to use these features without keras >= 3.0, found {keras.__version__=}") 
+
+from leap_net.keras.ltauNoAdd import LtauNoAdd
+from leap_net.keras.ltau import Ltau
+from leap_net.keras.resNetLayer import ResNetLayer
+
 __all__ = ["LtauNoAdd",
            "Ltau",
            "ResNetLayer",]